Add pmtiles, implement client side join with color mapping
