+++ popgetter-browser/web/package.json
@@ -55,6 +55,8 @@
     "chart.js": "^4.4.3",
     "chartjs-plugin-zoom": "^2.0.1",
     "comlink": "^4.4.1",
+    "d3-scale": "^4.0.2",
+    "d3-scale-chromatic": "^3.1.0",
     "duckdb": "^1.1.1",
     "flowbite-svelte": "^0.47.1",
     "flowbite-svelte-icons": "^1.6.2",
